Split into separate files, now to get it to compile...
